Module pg_extend::log

Implements macros for the PostgreSQL logging system.

For common log levels, convenient macros are implemented: trace!, debug!, log!, info!, notice!, warn!, error!, fatal!.

Other log levels are supported with the generic macro pg_log!. See the Level enum for all available log levels.

Note

Beware, log levels ERROR and higher also abort the current transaction. The PostgreSQL implementation uses exception handling with longjmp, which currently has unsafe side-effects.

Example

use pg_extend::{info, pg_log};
use pg_extend::log::Level;

info!("{} widgets frobnicated", 10);
pg_log!(Level::LogServerOnly, "Big brother is watching {}!", "you");

Rust log crate

The macro names make this mostly a drop-in replacement for the Rust log crate. However, there are differences:

Due to PostgreSQL behavior, log levels ERROR and higher log levels abort the current statement and transaction.

pg_extend macros do not support the optional target: argument.

In the log crate, the generic logging macro is called log!. However, we use that name as a specialized macro since PostgreSQL has a LOG log level.

Level enum contains Postgres-specific log levels; there is no Level::Trace for instance.

Enums

Level
Postgres logging Levels