Enum pg_extend::log::Level

pub enum Level {
    Debug5,
    Debug4,
    Debug3,
    Debug2,
    Debug1,
    Log,
    LogServerOnly,
    Info,
    Notice,
    Warning,
    Error,
    Fatal,
    Panic,
}

Postgres logging Levels

Note

Some of these levels effect the status of the connection and transaction in Postgres. Specifically, >= Error will cause the connection and transaction to fail and be reset.

Variants

Debug5

Debugging messages, in categories of 5 decreasing detail.

Debug4

Debugging messages, in categories of 4 decreasing detail.

Debug3

Debugging messages, in categories of 3 decreasing detail.

Debug2

Debugging messages, in categories of 2 decreasing detail.

Debug1

Debugging messages, in categories of 1 decreasing detail.

Log

Server operational messages; sent only to server log by default.

LogServerOnly

Same as LOG for server reporting, but never sent to client. CommError is an alias for this

Info

Messages specifically requested by user (eg VACUUM VERBOSE output); always sent to client regardless of client_min_messages, but by default not sent to server log.

Notice

Helpful messages to users about query operation; sent to client and not to server log by default.

Warning

Warnings. NOTICE is for expected messages like implicit sequence creation by SERIAL. WARNING is for unexpected messages.

Error

user error - abort transaction; return to known state

Fatal

fatal error - abort process

Panic

take down the other backends with me