Trait pg_extend::pg_alloc::RawPtr

pub trait RawPtr {
    type Target;
    unsafe fn from_raw(ptr: *mut Self::Target) -> Self;

    unsafe fn into_raw(self) -> *mut Self::Target;

    fn as_ptr(&self) -> *const Self::Target;
}

Types which implement this can be converted from pointers to their Rust type and vice versa.

Associated Types

type Target

Type to which the pointer is associated.

Loading content...
Required methods

unsafe fn from_raw(ptr: *mut Self::Target) -> Self

Instantiate the type from the pointer

Safety

Implementors should validate that all conversions into Rust wrapper type are within MemoryContexts

unsafe fn into_raw(self) -> *mut Self::Target

Consume this and return the pointer.

Safety

After calling into_raw there should be no other pointers to the data behind the pointer.

fn as_ptr(&self) -> *const Self::Target

Returns a pointer to this

Loading content...