Trait pg_extend::pg_datum::TryFromPgDatum

pub trait TryFromPgDatum<'s>: Sized {
    fn try_from<'mc>(
        memory_context: &'mc PgAllocator, 
        datum: PgDatum<'mc>
    ) -> Result<Self, &'static str>
    where
        Self: 's,
        'mc: 's;
}

A trait that allows for conversions between Postgres Datum types and Rust types.

Only Sized types, that fit in a single Datum, bool, u8 - u64 e.g. Nothing else is safe here.

Required methods

fn try_from<'mc>(
memory_context: &'mc PgAllocator,
datum: PgDatum<'mc>
) -> Result<Self, &'static str> where
Self: 's,
'mc: 's,

Attempt a conversion to from the Postgres data type into the Rust type

Loading content...