Enum pg_extend::pg_type::PgType

pub enum PgType {
    AbsoluteTime,
    BigInt,
    Int8,
    Boolean,
    GeoBox,
    ByteA,
    Char,
    Character,
    CommandId,
    Date,
    SmallInt,
    Int2,
    Int2Vector,
    Integer,
    Int4,
    Real,
    Float4,
    DoublePrecision,
    Float8,
    Interval,
    Lseg,
    Name,
    Oid,
    OidVector,
    Path,
    Point,
    RegProc,
    RelativeTime,
    Text,
    ItemPointer,
    Time,
    TimeWithTimeZone,
    Timestamp,
    TimeInterval,
    VarChar,
    Void,
    TransactionId,
}

See https://www.postgresql.org/docs/11/xfunc-c.html#XFUNC-C-TYPE-TABLE

TODO: it would be cool to share code with the sfackler/rust-postgres project though, that is converting from NetworkByte order, and this is all NativeByte order?

Variants

AbsoluteTime

abstime AbsoluteTime utils/nabstime.h

BigInt

bigint (int8) int64 postgres.h

Int8

bigint (int8) int64 postgres.h

Boolean

boolean bool postgres.h (maybe compiler built-in)

GeoBox

box BOX* utils/geo_decls.h

ByteA

bytea bytea* postgres.h

Char

"char" char (compiler built-in)

Character

character BpChar* postgres.h

CommandId

cid CommandId postgres.h

Date

date DateADT utils/date.h

SmallInt

smallint (int2) int16 postgres.h

Int2

smallint (int2) int16 postgres.h

Int2Vector

int2vector int2vector* postgres.h

Integer

integer (int4) int32 postgres.h

Int4

integer (int4) int32 postgres.h

Real

real (float4) float4* postgres.h

Float4

real (float4) float4* postgres.h

DoublePrecision

double precision (float8) float8* postgres.h

Float8

double precision (float8) float8* postgres.h

Interval

interval Interval* datatype/timestamp.h

Lseg

lseg LSEG* utils/geo_decls.h

Name

name Name postgres.h

Oid

oid Oid postgres.h

OidVector

oidvector oidvector* postgres.h

Path

path PATH* utils/geo_decls.h

Point

point POINT* utils/geo_decls.h

RegProc

regproc regproc postgres.h

RelativeTime

reltime RelativeTime utils/nabstime.h

Text

text text* postgres.h

ItemPointer

tid ItemPointer storage/itemptr.h

Time

time TimeADT utils/date.h

TimeWithTimeZone

time with time zone TimeTzADT utils/date.h

Timestamp

timestamp Timestamp* datatype/timestamp.h

TimeInterval

tinterval TimeInterval utils/nabstime.h

VarChar

varchar VarChar* postgres.h

Void

void

TransactionId

xid TransactionId postgres.h

Methods

impl PgType

pub fn from_rust<T: PgTypeInfo>() -> PgType

Return the PgType of the parameter's type

pub fn as_str(self, as_array: bool) -> &'static str

Return the string representation of this type

pub fn return_stmt(self, as_array: bool) -> String

Return the String to be used for the RETURNS statement in SQL